+++ src/ExtensibilityTools.cs
@@ -24,5 +24,6 @@ namespace MadsKristensen.ExtensibilityTools
         public const int cmdExportMoniker = 0x0004;
         public const int cmdSwatchesWindow = 0x0005;
         public const int cmdActivityLog = 0x0006;
+        public const int cmdVsipLogging = 0x0007;
     }
 }